Fix index offsets and base path
AI-generated code by Codex.

+++ SaintCoinach/IO/IIndexFile.cs
@@ -8,7 +8,7 @@ namespace SaintCoinach.IO {
     public interface IIndexFile : IEquatable<IIndexFile> {
         PackIdentifier PackId { get; }
         uint FileKey { get; }
-        uint Offset { get; }
+        long Offset { get; }
         byte DatFile { get; }
     }
 }
